add some funcs to utils
