+++ src/components/Section.js
@@ -1,0 +1,17 @@
+
+
+class Section {
+  constructor({items, renderer}, containerSelector) {
+    this._initialArray = items;
+    this._renderer = renderer;
+    this._container = document.querySelector(containerSelector);
+  }
+
+  renderItems() {
+    this._initialArray.forEach(item => this._renderer(item))
+  }
+
+  addItem(element) {
+    this._container.append(element);
+  }
+}
