+++ src/utils/utils.js
@@ -1,0 +1,25 @@
+import { ESC_KEYCODE } from "./constants";
+
+export const openModalWindow = (modalWindow) => {
+  modalWindow.classList.add('popup_is-opened');
+  document.addEventListener('keyup', handleEscUp);
+};
+
+
+export const closeModalWindow = (modalWindow) => {
+  modalWindow.classList.remove('popup_is-opened');
+  document.removeEventListener('keyup', handleEscUp);
+};
+
+
+export const handleEscUp = (evt) => {
+  evt.preventDefault();
+  isEscEvent(evt, closeModalWindow);
+};
+
+export const isEscEvent = (evt, action) => {
+  const activePopup = document.querySelector('.popup_is-opened');
+  if (evt.which === ESC_KEYCODE) {
+    action(activePopup);
+  }
+};
